Moved GetStorageProvider to namespace Orleans.Storage so it does not show up in intelisense unless one is using storage.
